move get-cache to tasks folder
